Refactor voucher and stock item timeline aggregation
Simplifies and optimizes the timeline aggregation logic in VoucherRepo and StockItemRepo. Removes redundant meta aggregation and summary calculations from VoucherRepo's viewTimeline, and updates StockItemRepo's viewTimeline to provide paginated and summarized stock movement data. Also removes unnecessary date filters and debug print statements for cleaner code.
